+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # SEOmatic Changelog
 
+## 1.1.51 - 2017.09.21
+### Added
+* Added support for `hreflang="x-default"`
+* Added `availabilty` to the default `Offer` Product JSON-LD
+
+### Changed
+* Fixed an issue where you could not change the Open Graph or Twitter `title` from a Twig template
+* Handle localized elements for breadcrumbs
+
 ## 1.1.50 - 2017.08.10
 ### Added
 * Add the ability to override the canonicalUrl from an SEOmatic Meta FieldType